requirements.txt
Flask
PyPDF2
python-dotenv
openai

